feat: moves server logic to package
