+++ go.mod
@@ -1,3 +1,5 @@
 module github.com/patrcoff/codecrafters-http-continued
 
 go 1.22.0
+
+replace github.com/patrcoff/codecrafters-http-continued/pkg/server => ../../pkg/server
